Trying to get openGL to work for graphing. I'm using win-ui again for development, since the qt ui designer is annoying. However, this is quite slow, as msvc2010 has no c++/cli intellisense. I'm creating the managed opengl graph control first, then exporting the opengl background logic to the qt control.
The win-ui project will be developed concurrently with qt-ui. win-ui can act as a
testbed for the project. It has native development, pre-built debugging symbols,
and is faster. I don't want to have to compile the new qt debug libraries again
for my laptop. Plus, it acts as a good review for managed code. The managed code
segments can be reused in C# and visual basic projects in the future.

Also worked on rtModel and rtRun. These will be the wrapper classes to the
actual library. Win-only at this point, but I'll also write the qt-equivalents
later.


git-svn-id: file:///mnt/svn/rtmath@31 1d52797c-35fc-48e8-8e80-2db72aa187e2