Bugfixes to quadrature! It now works properly with damatrices, I think. Tested against constant matrixop-derived damatrices, and the results agreed with hand calculations.
I had to disable the eval_cache in damatrix, for now. I need to reimplement it as
an unordered map and provide a hash function. Otherwise, my current implementation
randomly fails, which is not good. I'm glad that I gave it a master switch!

I turned off the netcdf code, also, as it turns out that there are no compilable c++
netcdf libraries on windows, which sucks. My own solution is insufficient for now, and needs
extensive testing.


git-svn-id: file:///mnt/svn/rtmath@187 1d52797c-35fc-48e8-8e80-2db72aa187e2