matrixop bug turned out to be a bad call to the set function. I was missing an argument. Oops. I converted most of the unsigned ints to size_t, as this is more appropriate.
Turned off msvc 2010 throw location mapping, as it was getting in the way of compilation. Also
turned off config.cpp, since it depends on boost filesystem, which is missing from the
virtual machine agravine (on galahad). It is a _large_ install, so it was easier to do this.

Matrixop needs repivoting tested. Shall create appropriate tests for linux, as agravine needs a
restart to update itself. Then, I need to use the upper triangular matrix in repivoting operations.
I also have to check if repivot can repivot a second matrix. Will be used in inverse and eigen* calculations.


git-svn-id: file:///mnt/svn/rtmath@218 1d52797c-35fc-48e8-8e80-2db72aa187e2