All of the recursive polynomials (except for associated legendre) have been added and the tests have been constructed and passed. The associated legendre polynomials are more complex and will not be treated unless I have need of them. They will have a different class structure than the rest.
For polynomial comparisons, a new approxEq method has been added, since the recursive polynomial tests kept failing due to differences in the coefficients of one part in 10^-12 or less. approxEq validates the coefficients to 0.0001%. approxEq for a double works similarly. Please note that approxEq only validates the coefficients of the polynomial, not the domain/range of approximately equal values.

Makefile modified to add polynomials/*.cpp files.



git-svn-id: file:///mnt/svn/rtmath@281 1d52797c-35fc-48e8-8e80-2db72aa187e2