I've gutted absorb, lbl and atmos. The current implementation was not good, as any absorption calculations were too lbl-dependent, and lbl was slow and wasn't quite right, anyways. This will break my transmittance programs until fixed. However, it will allow for better atmospheric calculations, and it takes the wv continuum properly into account as well. I need to check the conversions to make sure that collision-induced absorption is in dB and not nepers.
Next will be to redo the transmittance programs and to check absorbance. Then, I need to redo
the rest of the atmosphere. Then, I need to reimplement doubling-adding / check that it is
correct. Then, I need to fix ddscat's issue with orientation. After that, I have to verify
that all of its calculations (including emission vectors) are correct and run it over the
frequency domain. That data will make its way into my poster. Lots to do.....


git-svn-id: file:///mnt/svn/rtmath@310 1d52797c-35fc-48e8-8e80-2db72aa187e2