Important work on rtmath::config. I implemented shared_from this inheritance, so I don't have to use the create(...) functions. I can instead directly call the constructur, which is publicly-accessable.
config also now fully implements ostream and istream overrides, so it works with operator<< and >>.
The getPtr function is const, but the return is not. The cast is rather interesting, and I think that breaking const-ness is
required in this situation. Otherwise, I'd need a major rewrite.

Symlinks have been envisioned, but the support is nonexistent. How will I handle file includes? Also, how will I read/write them?
I want to implement them to allow Includes in my config outputs/saves, but I need to think of an implementation.
Possibly, I'd use strong_ptrs to keep the included sections in the memory tree, and add a guid member to avoid crossing the
boundary to the symlink write area. Upon reaching this boundary, an Include line may be generated, as specified in the
symlinked segment.

TODO: add a file source anyways, to allow for updating the config. This would leverage ostream anyways.
TODO: merge Include and IncludeIfExists.
TODO: symlinks
TODO: better pathing. with symlinks, either disable searching across boundaries or try to avoid infinite loop situations.
 Can avoid infinite loops by looking at sets of already-crossed guids.


git-svn-id: file:///mnt/svn/rtmath@451 1d52797c-35fc-48e8-8e80-2db72aa187e2