Project builds on gcc again. Need to work more on source writing. CMakeLists updated to add install targets for all apps and libraries. Need to add appropriate target for test code.
Apps now link to dynamic rtmath. LD_LIBRARY_PATH and LIBRARY_PATH need to 
reflect where the libraries are. This in not to be covered by this build 
system, as the user may specify the destination prefixes. See the standard 
module file on solar for hints. May eventually include a file like this 
if I produce packaged builds anytime.



git-svn-id: file:///mnt/svn/rtmath@453 1d52797c-35fc-48e8-8e80-2db72aa187e2