First, ROOT_functions.cpp and .h added, to provide in-library some common ROOT routines. A mock-up of a y-axis flipping function was created, but it doesn't work at all. There are too many features from the initial axis that need to be cloned, and the optiong between TAxis and TGaxis do not match up entirely. So, it's on the backburner...
Next, matrixop::print now takes a stream as an argument.

shape loading routines debugged. They now work, but the statistics should be places in a separate class, with a constructor that takes
the loaded shape. The shape, btw, should not be placed in a matrixop because it is sparse. Coords now debugged so matrixop conversions
work as expected. Sets of these can act as sparse matrix points, but it really is awkward. I need to make a better notation for them.

rtmath-shape project added to allow for shape manipulations and statistics calculations.

config debugged. Major bug was introduced (and now is fixed) where rtconf loading returned the wrong pointer (a new object's parent
instead of the object), which failed when loading the root node.

atmosplot now takes -h option for help generation.


git-svn-id: file:///mnt/svn/rtmath@467 1d52797c-35fc-48e8-8e80-2db72aa187e2