from_file implementation slowly progressing. need to adapt shapefile class to allow for rescaling to aspect ratios and other quantities. shapefilestats would be helpful here.
shapefilestats will be able to figure out the rotation orientation energy function and
provide reorientation information. it will provide aspect ratios in a variety of manners,
based on mean, rms and absolute x,y,z dimensions. it should then be able to modify a shapefile
header to accomplish these changes.


git-svn-id: file:///mnt/svn/rtmath@626 1d52797c-35fc-48e8-8e80-2db72aa187e2