More tests written. ddPar now can serialize and load properly (bug fixed with extraneous it++) in scaPlanes. The standard read and write functions work as well. ddParSimple now also has comparison and strict ordering.
Work done on serialization code to allow for compression / decompression. Most of the
stuff should work, but the code to handle boost::iostreams directly is clearly
incorrect and does not compile. I need to read up on this to make sure that
my sources and sinks are correct.

After this, I'll test compression to ensure that it works, and might alter read and write
such that compression is not done at the file io level. Instead, the method will be
passed at the stream level.


git-svn-id: file:///mnt/svn/rtmath@634 1d52797c-35fc-48e8-8e80-2db72aa187e2