Seriously done with ddPar class for a while!
Worked on testing set functions which revealed several code flaws. Redid accessors as
template functions to help in debugging.
New objects that are plural (containers) need to have a separate resize function to help
sizing, especially when a key is not loaded by default. This was implemented, and
all of the constructors had to be updated accordingly.

Also, ddPar read was reading the first (useless) line, which was suppressed.

Traditional functions are called by the enhanced interface, so they need not be tested
separately. They've always worked, anyways.

So, ddPar passes all of the tests!!!!!

stdafx updated to finally get rid of the lexical_cast warnings.


git-svn-id: file:///mnt/svn/rtmath@642 1d52797c-35fc-48e8-8e80-2db72aa187e2