splitString GETOBJKEY() removed, as aliases work.
In test_ddparGenerator, cannot compare two diel entries, as ordering in serialization is
not fixed (map based on string and shared_ptr addresses). So, only use one.
Check if other constraints also upset this, or provide operator<, operator== and operator!=.

GETOBJKEY in shapeModifiable retained, as vertex and graph saving still needs to be done.
This requires depGraph modifications, but it should be trivial.

Will change this and do generator tests next.


git-svn-id: file:///mnt/svn/rtmath@647 1d52797c-35fc-48e8-8e80-2db72aa187e2