shape-main worked on headers here, in stdafx and rtmath.h to get rid of warnings, and to get rid of some imports of ROOT (which were messing up msvc error messages with boost accumulator bugs and the new matrixop template functions).
matrixop now has to<> and from<> operators and overrides to allow for io to other classes!
test constructed and validated.

ddpar and ddpargenerator work for stability.

shapefilestats now has base class, and it can be serialized. moments are calculated for starters,
and other quants will be saved too. no longer using matrixop pointers, just raw matrixops which
get initialized in the base.


git-svn-id: file:///mnt/svn/rtmath@666 1d52797c-35fc-48e8-8e80-2db72aa187e2