Basic serialization has been implemented for tmatrix. This will not be the scheme used by rtmath, however! Converters will be implemented for this.
Apps have been loaded. tmatrix-batch-runs is in a uncompliable state due to the tmatrix
changes, and it is in need of a new input and output mechanism. Will instead rewrite the code
to have the dispatcher handle all of the file output. Moreso, the batch program will need to
take another parameter specifying the output file. Requires more elaborate worker /
dispatcher return communication.


git-svn-id: file:///mnt/svn/rtmath@704 1d52797c-35fc-48e8-8e80-2db72aa187e2