shapestatsviews now compiles and is used in one of the programs. tmData class created that gives a header and links the stats to the tmatrix converter. tmatrix-convert now fully implemented for the isotropic case. need to finish the stat unit check and redo.
ddscattmatrix needs to be split into two derived classes, depending on if f is known.
the other class will just hold P. needed for ddavg import.

need to switch std to boost shared_ptrs and need to serialize the rest of the ddscat code.

then, q and backscatter can be calculated and compared.


git-svn-id: file:///mnt/svn/rtmath@741 1d52797c-35fc-48e8-8e80-2db72aa187e2