mie program extended to allow my pf code to calculate the phase function based on the s inputs. I've validated that swapping the tmatrix s real and imaginary values validates the pf predicted by mie theory.
I have lots of literature showing that my mie S and P11 are correct.
My other mie P values are off... Will link into the tmatrix generator for this validation.

Is it a difference in notation between mie and tmatrix? I don't think so, as I have checked this.

I will split the tmatrix S / P conversion next and implement it in a separate linkable function. This should give me the correct phase matrices given a correct S input. After this comes the decision concerning what to do with the S values.


Also, the mttest program now handles the input angle to radian conversion for the mie code correctly. The S values now are a perfect matchup for spheres with re and im swapped.



git-svn-id: file:///mnt/svn/rtmath@775 1d52797c-35fc-48e8-8e80-2db72aa187e2