ddscat-data created. This program will interconvert divverent ddscat output types, and acts to demonstrate  that ddscat files can be read properly.
It reads everything but the headers (curently) for fml, sca and avg files. It gets the stat tables correct.

It can display F, Mueller and stat tables. 

Indicates that my genMuellerMatrix code does not work for ddscat, as expected. My genMueller code is 
currently the tmatrix implementation. I need to switch to the e^{-ikx} convention (perhaps pulling from 
an old revision and splitting into two separate functions).

ddScattMatrix and ddOutputSingle have undergone changes to ensure that this all works, of course.

I have preliminary write support, but the reads need to process the header properly and need to fill 
in any missing, calculable values.



git-svn-id: file:///mnt/svn/rtmath@801 1d52797c-35fc-48e8-8e80-2db72aa187e2