Created the application liu-regenerate-runs, which is designed to regenerate the shape files that Liu uses for his database. This reads in the file ids, determines which type of shape needs generation along with its relevant dipole spacing, and produces the relevant shape files. The dipole spacings are placed in a tsv file for reference.
It is planned that the tmatrix conversion code will be changed to enable reading of this file,
matching its contents to any relevant raw or stats files, and performing the run with the
appropriate parameters.

Modified ddOutputSingle to provide TARGET info. It can now provide a raw form of any of the
header lines.


git-svn-id: file:///mnt/svn/rtmath@876 1d52797c-35fc-48e8-8e80-2db72aa187e2