Mean is calculated correctly (or very close - x0 vs xd).
Dipole spacing has an issue - the shape volume does not correspond to the dipole volume. For example, in a aeff 5 standard hex plate, V=523.599, but it has 4257 dipoles, with Vd ~ 1.0528. Problem persists for d > 1.

Likewise, switching from d=1 to d=2 presents Nd1 = 4257 and Nd2 = 845. 4257 / 845 = 5.038! The shape is correct, but I need a consistent dipole relation.



git-svn-id: file:///mnt/svn/rtmath@1006 1d52797c-35fc-48e8-8e80-2db72aa187e2