ConfigSegment still has some issues with finding the right key. Made todos visible in Doxygen.
Shapefile modified to properly write compressed file names. Still need to handle compressed reads. Since the number of points are provided, this shouldn't be much of an issue. It just involves changing the string function to take a stream.

shape-hash works. It finds keys correctly in configSegment. It writes compressed output. It appends stats. It writes debug output.



git-svn-id: file:///mnt/svn/rtmath@1055 1d52797c-35fc-48e8-8e80-2db72aa187e2