Building on the serialization fixes earlier, I added support for multiple serialization types. I had to correct hash.cpp, ddOutputSingle and shapestats to take advantage of this.
The shapestats hash function now writes the serialized extension (.xml or .ts) in the filenames. ddOutputSingle has moved the read and write serialized sections into their own 'if' blocks to take advantage of the different formats available.

I'm now using the Ryan_Serialization::known_format function to get the formats available from built-in maps or from the console.

I have to update data.cpp to support the Ryan_Serialization program_options functions.

ddOutputSingle has a bug when reading a written .avg file, as the refractive index field does not align in the same way as getM is wont to do. I will correct this in the morning.
I still also have to check over the shapestats calculations, as I am getting nans and invalid numbers.

Once these are done, I can move on to the weights and finish ddOutput and implement the ddOutputGenerators.

git-svn-id: file:///mnt/svn/rtmath@1089 1d52797c-35fc-48e8-8e80-2db72aa187e2