Much of the Von Mises-Fisher 3d orientation has been added. The CDF and PDF code needs to be fixed. Mus and xs should have 3 elements each. It would be helpful to include a converter function.
All of the classes in ddWeights now have accessor functions to read the intervals used for weighting. This will be better than rying to search for the pivot point constantly.

orientation-weights-3d started, but I need to finish the vMF code first. I also need to add an option for superpositions of vMF weights for opposite theta sides. Symmetry is required to support rotation of a flake by 180 degrees. This might entail another orientation weighting class (easy).


git-svn-id: file:///mnt/svn/rtmath@1119 1d52797c-35fc-48e8-8e80-2db72aa187e2