The reading routine is faster than before. I switched from using lexical_cast in readF and readMueller in favor of using my own parsers, m_atof and a Boost Spirit-based parser in an anonymous namespace. Read times are improved by about a factor of six.
For shape files, I turned off an extraneous hashing operation. The hash takes the longest time to compute because the shape has to be written out using streams. I can't hash the input directly because of the streams used for compression. I'll look into it. Also, the shape loads were slightly sped up by using threads for stat calculation at the expense of code clarity. If I switch from streams to a large block of memory, the older svn routines were already quite fast. I'd be able to parallelize the getline operation easily.

Added debug::getConcurrentThreadsSupported that gives the optimal number of threads on a system. I need to expand this with os-specific code.


git-svn-id: file:///mnt/svn/rtmath@1128 1d52797c-35fc-48e8-8e80-2db72aa187e2