Lots of work on implementing Voronoi diagram generation in a manner that is useful for shape analysis. The VoronoiDiagram class was implemented, along with methods for finding candidate convex hull points and calculation of depth from 'surface'. The surface depth function needs to be tested and somehow combined with shapefile in a program to view the output.
hulls.h is aging fast. VTK code needs to be spun off into a plugin, but with the proviso that it is used in the convex hull calculation. Is this calculation necessary outside of shape stat calculation? It does reduce the number of points needed for the max diameter calculation.

Modified depGraph to also return node depth. Need to check rtmath_ddscat for any impacted functions (will show up on next rebuild).

Removed shapefile built-in bov writing code. bov files still unreadable in Visit, so there must be a small error.


git-svn-id: file:///mnt/svn/rtmath@1221 1d52797c-35fc-48e8-8e80-2db72aa187e2