Voronoi stuff becoming frustrating. Generates output, but all of the "order" parameters are either zero or one.
Need to switch to unordered_sets for fast reservation / allocation / deallocation. However, it is the deallocation that kills performance.
May need either boost shared memory or may need to abandon the set implementation for pure arrays.

Once this routine is sped up and fixed, will implement the true surface routine. Will also see if the voronoi container accesses can be parallelized and precached.


git-svn-id: file:///mnt/svn/rtmath@1224 1d52797c-35fc-48e8-8e80-2db72aa187e2