Fixed all known hdf write issues. The data were screwed up because the Eigen internal matrix was in column-major order, and the hdf5 library writes in row-major order. The mean calculation of the shapefile is messed up again. Will fix since it affects latticePtsNorm.
It is now time to work on hdf5 output of the relevent shape statistics (and stat vertification).
I want to be able to show Liu an output plot of Q... vs volume, with contours of SA/V ratio, for multiple flake types.
This requires being able to read in ddOutput avg results and to perform stat calculation with the shapes. Also needs an updated shapestatsviews routine to handle appropriate units.

Also need to check / redo tmatrix output. Need to give plots of same volume spheroids for different aspect ratios. Contours maybe?


git-svn-id: file:///mnt/svn/rtmath@1245 1d52797c-35fc-48e8-8e80-2db72aa187e2