Copied (not moved) the tsv export code in ddscat-output-extract and ddscat-output-extract-orientations into the tsv plugin. I have to adapt the applications, which are now redundant with each other.
Began planning the hdf5 output structure, as this file format will be used as the primary data storage format for future modeling initiatives.

Redid rtmath-apps selections to ease development.

ddOutput can now guess the temperature for a loaded set of data from the frequency and refractive index. Should be to within 1 K.


git-svn-id: file:///mnt/svn/rtmath@1306 1d52797c-35fc-48e8-8e80-2db72aa187e2