Implemented a very basic version of the radar-relations program. It has no support for outputting / exporting files, except for the three debugging output files. These output files will all be placed into the tsv writer later.
It lacks an implementation of a particle size-frequency distribution, and all particles have the same canting angle distribution. Finally, it still lacks the ability to find the PE minimum, so any zero angle must be specified manually (shapestats version takes far too long).

I'll do some basic runs with this version, and will try to get it to run without error on Linux.


git-svn-id: file:///mnt/svn/rtmath@1308 1d52797c-35fc-48e8-8e80-2db72aa187e2