Probably won't fix the compression issue for now. Need to also remember to add compression support to standard text writes as well.
Will apply the new interface to all of the existing classes tomorrow. There will be some difficulties with the hash implementation, which will be temporarily disabled.
Also need to modify the read interface, adding an optional member to read in data filling a criteria. If no limiting criteria are chosen,
then all data in a file (i.e. hdf5) will be imported to a vector.


git-svn-id: file:///mnt/svn/rtmath@1354 1d52797c-35fc-48e8-8e80-2db72aa187e2