All of the libraries are using the new io code! Most apps will need to be updated. The regenerate program needs special attention, as it only reads shapefile headers, and this function now using strings of input, not filenames.
From here, I need to modify the io read and write functions to allows multiple inputs and outputs in one operation. For example, an operation that searches through an hdf5 file and only retrieves results within a frequency range. The selector will come later, but I first want to be able to pull everything into a vector first. 

After this, I'll take the radar-relations program and attempt to prescribe a given particle size / number density relationship.



git-svn-id: file:///mnt/svn/rtmath@1356 1d52797c-35fc-48e8-8e80-2db72aa187e2