Lots of splitting ddOutputSingle into smaller files. I no longer run into the object size limitation, though the code does take a very long time to compile.
In the interests of performance, I need to change how files are read.
For memory alignment, the cross-section objects need an optional constructor argument to relocate to a ddOutput container-provided storage.

The ddScattMatrices container should be changed to a vector of the actual objects, as they are always linked to a single ddOutputSingle. No ordering will be required, speeding up execution and placement times.

There should be an easy facility to link orientation weights with the actual ddOutputSingle entries. For future research, it needs to be fast.


git-svn-id: file:///mnt/svn/rtmath@1357 1d52797c-35fc-48e8-8e80-2db72aa187e2