More progress on ddOutput modifications. The finalize function is introduced for the (now one) generator to call upon ddOutput creation. It needs to be filled in / updated to use the most current definitions.
Also, I have to check to see that the ddOriData class is updating the integer variables relating to number of fml matrices (and to see if a rotation should be weighted). Will remove numSCAs from the definitions.

Finally, need to properly import the number of orientations that an avg file uses (can pull from a par file) for use in regenerating a weighting table.


git-svn-id: file:///mnt/svn/rtmath@1429 1d52797c-35fc-48e8-8e80-2db72aa187e2