Many changes that build but are untested. qhull plugin added. For now, just provides convexHull functionality. Should speed up some calculations.
qhull is used by default with the tsv-export of a shape structure. The goal of this is to report a shape structure with a distance from the center and a max_radius feature.

The max_radius feature can be rewritten such that the furthest point is scaled to 1. This would be a two-pass system, and should work. Probably will shift to this to improve overall file write speed.

This is all done to achieve the next step of my research. I want to construct histograms of radial mass distributions at different particle sizes to see if they are the same. I can calculate the confidence bounds and use this to create a generalized mass distribution function.

Some todos from before:
- The silo output works, but need to add a supplementary matrix where the convoluted volume fraction is reported.
- Need to add logic that determines a convolution size from the wavelength.
 - Statistical testing is needed for this.



git-svn-id: file:///mnt/svn/rtmath@1811 1d52797c-35fc-48e8-8e80-2db72aa187e2