Extended convolution to handle edges correctly.
Convolution code may yield slightly different outputs depending on whether convolution is in same program invocation as output, or if they are chained. The slight difference in X0 changes the radii and the shape hash. Different numbers of output points may result. The overall effect is quite minor, so I'll ignore it for now.

Extended silo and tsv output to add a normed dielectric value, which is useful when considering convolution across a certain radius. Now, I can immediately get a local estimate of the volume fraction. May be able to use this in subsampling DDA!

Extended shape-extract to specify export type, so it now can incorporate tsv output. No need to chain with shape-output.

shapefile::convolute now takes the convolution radius as a parameter. Needed for the edge resizing.

tsv and silo export plugins now take the "dielScalingFactor" option when specifying the normalization on the 'dielectric'. Actual value is divided by this (usually expressed as a volume). silo writes the appropriate data on the existing mesh; tsv just adds a column.



git-svn-id: file:///mnt/svn/rtmath@1813 1d52797c-35fc-48e8-8e80-2db72aa187e2